+++ Source/Modules/Catalog/Reflection/Gapotchenko.FX.Reflection.Loader/Polyfills/FileSystem.cs
@@ -1,4 +1,6 @@
-﻿namespace Gapotchenko.FX.Reflection.Loader.Polyfills;
+using System.Runtime.InteropServices;
+
+namespace Gapotchenko.FX.Reflection.Loader.Polyfills;
 
 static class FileSystem
 {
@@ -7,5 +9,30 @@ static class FileSystem
             StringComparer.InvariantCulture :
             StringComparer.InvariantCultureIgnoreCase;
 
-    public static bool IsCaseSensitive { get; } = Environment.OSVersion.Platform == PlatformID.Unix;
+    public static bool IsCaseSensitive { get; } = IsCaseSensitiveCore();
+
+    static bool IsCaseSensitiveCore()
+    {
+        if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows) ||
+            RuntimeInformation.IsOSPlatform(OSPlatform.OSX))
+        {
+            return false;
+        }
+        else if (RuntimeInformation.IsOSPlatform(OSPlatform.Linux)
+#if NETCOREAPP3_0_OR_GREATER
+            || RuntimeInformation.IsOSPlatform(OSPlatform.FreeBSD)
+#endif
+            )
+        {
+            return true;
+        }
+        else
+        {
+            // A graceful fallback.
+            if (Environment.OSVersion.Platform == PlatformID.Unix)
+                return true;
+            else
+                return false; // a safer default
+        }
+    }
 }